linux mac type issue fixed

+++ src/ft_time.h
@@ -6,7 +6,7 @@
 /*   By: rmeiboom <marvin@codam.nl>                   +#+                     */
 /*                                                   +#+                      */
 /*   Created: 2021/12/06 11:20:47 by rmeiboom      #+#    #+#                 */
-/*   Updated: 2022/01/11 14:13:33 by rmeiboom      ########   odam.nl         */
+/*   Updated: 2022/01/11 19:59:36 by rmeiboom      ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,7 +15,7 @@
 
 # include "defines.h"
 
-typedef u_int64_t	t_time_ms;
+typedef unsigned long long	t_time_ms;
 
 t_time_ms	get_time_ms(void);
 void		sleep_ms(t_time_ms sleep_ms);
